refactor scalar decimal module
